requirements.txt
PyYAML>=6.0
psutil>=5.9.0
phue>=1.1